+++ pages/api/v1/auth/[provider].ts
@@ -1,15 +1,15 @@
 import nextConnect from 'next-connect';
 
-import { Provider } from '../../../src/client/context/AuthContext/AuthContext';
-import { ENABLED_PROVIDER } from '../../../src/constants';
+import { Provider } from '../../../../src/client/context/AuthContext/AuthContext';
+import { ENABLED_PROVIDER } from '../../../../src/constants';
 import {
   authMiddleware,
   promisifyAuthentication,
-} from '../../../src/server/auth/middlewares';
+} from '../../../../src/server/auth/middlewares';
 import {
   NOT_FOUND,
   INTERNAL_SERVER_ERROR,
-} from '../../../src/utils/statusCodes';
+} from '../../../../src/utils/statusCodes';
 
 export default nextConnect()
   .use(authMiddleware)
