+++ pages/api/v1/auth/logout.ts
@@ -1,7 +1,7 @@
 import { NextApiRequest, NextApiResponse } from 'next';
 
-import { removeTokenCookie } from '../../../src/server/auth/cookie';
-import { OK } from '../../../src/utils/statusCodes';
+import { removeTokenCookie } from '../../../../src/server/auth/cookie';
+import { OK } from '../../../../src/utils/statusCodes';
 
 export default function handler(_: NextApiRequest, res: NextApiResponse) {
   removeTokenCookie(res);
